indexers should be ints since they relied on java constructs

+++ src/main/java/org/bytedeco/javacpp/BytePointer.java
@@ -90,7 +90,7 @@ public class BytePointer extends Pointer {
      *
      * @param size the number of {@code signed char} elements to allocate
      */
-    public BytePointer(int size) {
+    public BytePointer(long size) {
         try {
             allocateArray(size);
         } catch (UnsatisfiedLinkError e) {
@@ -101,7 +101,7 @@ public class BytePointer extends Pointer {
     public BytePointer() { }
     /** @see Pointer#Pointer(Pointer) */
     public BytePointer(Pointer p) { super(p); }
-    private native void allocateArray(int size);
+    private native void allocateArray(long size);
 
     /** @see Pointer#position(long) */
     @Override public BytePointer position(long position) {
@@ -183,7 +183,7 @@ public class BytePointer extends Pointer {
     /** @return {@code get(0)} */
     public byte get() { return get(0); }
     /** @return the i-th {@code byte} value of a native array */
-    public native byte get(int i);
+    public native byte get(long i);
     /** @return {@code put(0, b)} */
     public BytePointer put(byte b) { return put(0, b); }
     /**
@@ -193,7 +193,7 @@ public class BytePointer extends Pointer {
      * @param b the {@code byte} value to copy
      * @return this
      */
-    public native BytePointer put(int i, byte b);
+    public native BytePointer put(long i, byte b);
 
     /** @return {@code get(array, 0, array.length)} */
     public BytePointer get(byte[] array) { return get(array, 0, array.length); }
